Policies

+++ app/Http/Controllers/ThreadController.php
@@ -113,6 +113,11 @@ class ThreadController extends Controller
      */
     public function destroy($channel, Thread $thread)
     {
+        $this->authorize('update', $thread);
+
+        if($thread->user_id != auth()->id()){
+            abort(403, 'You do not have permission to do this.');
+        }
         $thread->replies()->delete();
         $thread->delete();
